What I've done... Added bluetooth server/client socket Devices can communicate
Problems to work on...
In order for devices to communicate I have to manually click on
the paired device i want to communicate with on only one of the devices, the other waits in the main activity and listens for the client socket, which starts up after the messenger activity is opened, to find it. After around 3 seconds the devices are able to communicate and the server side device can enter the messenger activity and freely communicate. There are a host of issues that come with this problem and I'm trying to rearrange the way i start the client and server socket connections.